Factor MockThemeLoader to separate file
