+++ browser/test/Services/TokenColorsTests.ts
@@ -6,38 +6,9 @@ import * as assert from "assert"
 
 import { TokenColors } from "./../../src/Services/TokenColors"
 
-import { MockConfiguration } from "./../Mocks"
-// import * as TestHelpers from "./../TestHelpers"
+import { MockConfiguration, MockThemeLoader } from "./../Mocks"
 
-import { IThemeContribution } from "./../../src/Plugins/Api/Capabilities"
-import {
-    DefaultTheme,
-    IThemeLoader,
-    ThemeManager,
-    IThemeMetadata,
-} from "./../../src/Services/Themes"
-
-export class MockThemeLoader implements IThemeLoader {
-    private _nameToTheme: { [key: string]: IThemeMetadata } = {}
-
-    public getAllThemes(): Promise<IThemeContribution[]> {
-        const themeContributions = Object.keys(this._nameToTheme).map(
-            (name): IThemeContribution => ({
-                name,
-                path: null,
-            }),
-        )
-        return Promise.resolve(themeContributions)
-    }
-
-    public getThemeByName(name: string): Promise<IThemeMetadata> {
-        return Promise.resolve(this._nameToTheme[name])
-    }
-
-    public addTheme(name: string, theme: IThemeMetadata): void {
-        this._nameToTheme[name] = theme
-    }
-}
+import { DefaultTheme, ThemeManager } from "./../../src/Services/Themes"
 
 describe("TokenColors", () => {
     let mockConfiguration: MockConfiguration
@@ -62,13 +33,23 @@ describe("TokenColors", () => {
         assert.strictEqual(hitCount, 1, "Validate onTokenColorsChanged was fired")
     })
 
-    it("setting token configuration value triggers onTokenColorsChanged event", () => {
+    it("if configuration is updated, but 'editor.tokenColors' isn't there, onTokenColorsChanged should not be triggered", () => {
         const tokenColors = new TokenColors(mockConfiguration as any, themeManager)
 
         let hitCount = 0
         tokenColors.onTokenColorsChanged.subscribe(() => hitCount++)
 
         mockConfiguration.simulateConfigurationChangedEvent({ someConfigValue: 2 })
+        assert.strictEqual(hitCount, 0, "Validate onTokenColorsChanged was fired")
+    })
+
+    it("if 'editor.tokenColors' is updated from the config, onTokenColorsChanged event should be triggered", () => {
+        const tokenColors = new TokenColors(mockConfiguration as any, themeManager)
+
+        let hitCount = 0
+        tokenColors.onTokenColorsChanged.subscribe(() => hitCount++)
+
+        mockConfiguration.simulateConfigurationChangedEvent({ "editor.tokenColors": [] })
         assert.strictEqual(hitCount, 1, "Validate onTokenColorsChanged was fired")
     })
 })
